+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/DocumentField/hooks/useDocumentUpload/helpers/get-file-or-fileid-from-documents-list/get-file-or-fileid-from-documents-list.ts
@@ -1,7 +1,7 @@
 import { IFormElement } from '@/components/organisms/Form/DynamicForm/types';
 import get from 'lodash/get';
 import { IDocumentFieldParams } from '../../../../DocumentField';
-import { composePathToFileId } from '../compose-path-to-file-id';
+import { composePathToDocumentPageProperty } from '../compose-path-to-document-page-property';
 
 export const getFileOrFileIdFromDocumentsList = (
   documentsList: Array<IDocumentFieldParams['template']> = [],
@@ -15,7 +15,7 @@ export const getFileOrFileIdFromDocumentsList = (
     return undefined;
   }
 
-  const filePath = composePathToFileId(documentIndex, pageProperty, pageIndex);
+  const filePath = composePathToDocumentPageProperty(documentIndex, pageProperty, pageIndex);
   const fileOrFileId = get(documentsList, filePath, undefined);
 
   return fileOrFileId;