+++ apps/workflows-dashboard/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ballerine/workflows-dashboard
 
+## 0.2.37
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.82
+  - @ballerine/ui@0.7.117
+
+## 0.2.36
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.81
+  - @ballerine/ui@0.7.116
+
+## 0.2.35
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.80
+  - @ballerine/ui@0.7.115
+
 ## 0.2.34
 
 ### Patch Changes