+++ packages/rules-engine/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/rules-engine-lib
 
+## 0.5.36
+
+### Patch Changes
+
+- version bump
+
+## 0.5.35
+
+### Patch Changes
+
+- bump
+
+## 0.5.34
+
+### Patch Changes
+
+- version bump
+
 ## 0.5.33
 
 ### Patch Changes