+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.103
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.82
+  - @ballerine/workflow-core@0.6.103
+
+## 0.6.102
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.81
+  - @ballerine/workflow-core@0.6.102
+
+## 0.6.101
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.80
+  - @ballerine/workflow-core@0.6.101
+
 ## 0.6.100
 
 ### Patch Changes